Move shared types to shared/types. Update build accordingly.
