+++ package.json
@@ -29,7 +29,7 @@
   "activationEvents": [],
   "preview": true,
   "icon": "byo_LAD.png",
-  "main": "./out/extension.js",
+  "main": "./out/src/extension.js",
   "contributes": {
     "configuration": {
       "title": "byoLAD",
