+++ shared/types/Chat.ts
@@ -1,0 +1,12 @@
+import { ChatMessage } from "./ChatMessage";
+
+export interface Chat {
+  id: number;
+  name: string;
+  messages: ChatMessage[];
+  /**
+   * Base contextual instruction to inform the LLM of its purpose/goal/role
+   */
+  contextInstruction?: string;
+  title: string;
+}
