+++ shared/types/ChatRole.ts
@@ -1,0 +1,5 @@
+export enum ChatRole {
+  User = "user",
+  System = "system",
+  Assistant = "assistant",
+}
