+++ shared/types/index.ts
@@ -1,0 +1,4 @@
+export * from "./Chat";
+export * from "./ChatMessage";
+export * from "./ChatRole";
+export * from "./MessageBlock";
