+++ src/commands/getReviewCodeCommand.ts
@@ -4,8 +4,7 @@ import { SettingsProvider } from "../helpers/SettingsProvider";
 import { ChatManager } from "../Chat/ChatManager";
 import { ChatWebviewProvider } from "../providers/ChatViewProvider";
 import { insertMessage } from "../helpers/insertMessage";
-import { TextBlock } from "../ChatModel/ChatModel";
-
+import { TextBlock } from "../../shared/types";
 /**
  * Command to review the selected code (or whole file if no selection) in a chat.
  * Sends the selection/file and the user's configured prompt as a chat message.
