+++ src/test/suite/extension.test.ts
@@ -3,7 +3,7 @@ import * as assert from "assert";
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
 import * as vscode from "vscode";
-import { CodeBlock, MessageBlock, TextBlock } from "../../ChatModel/ChatModel";
+import { CodeBlock, MessageBlock, TextBlock } from "../../../shared/types";
 import { stringToMessageBlocks } from "../../Chat/messageBlockHelpers";
 
 suite("Extension Test Suite", () => {
