+++ tsconfig.json
@@ -5,8 +5,8 @@
     "outDir": "out",
     "lib": ["es6", "dom"],
     "sourceMap": true,
-    "rootDir": "src",
     "strict": true
   },
+  "include": ["./src/**/*.ts", "../shared/**/*.ts"],
   "exclude": ["node_modules", ".vscode-test", "webview-ui"]
 }
