+++ webview-ui/src/types.ts
@@ -1,13 +1,3 @@
-export enum MessageType {
-  User,
-  AI,
-}
-
-export interface Message {
-  type: MessageType;
-  message: string;
-}
-
 export interface ImagePaths {
   byoLadCircleImageUri: string;
 }
